% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{cluster_markers}
\alias{cluster_markers}
\title{Identify markers in each cell population}
\usage{
cluster_markers(
  element,
  clusting_results,
  assay,
  features = NULL,
  test_using = "bimod",
  thresh_pc = 0.25,
  thresh_fc = 0.25,
  thresh_p = 0.01
)
}
\arguments{
\item{element}{Eigenexpression matrix}

\item{clusting_results}{UMAP/t-SNE results}

\item{assay}{Define the assay to be displayed, e.g., assay = “RNA”}

\item{features}{Eigenvector}

\item{test_using}{Test to be used (“bimod” or “wilcox”)}

\item{thresh_pc}{Threshold}

\item{thresh_fc}{Threshold}

\item{thresh_p}{Threshold}
}
\description{
Identify markers in each cell population
}
