% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{DifferentialLRT}
\alias{DifferentialLRT}
\title{Likelihood ratio test}
\usage{
DifferentialLRT(x, y, xmin = 0)
}
\arguments{
\item{x}{Data}

\item{y}{Data}

\item{xmin}{parameter}
}
\description{
Likelihood ratio test
}
