% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{feature_score_visual}
\alias{feature_score_visual}
\title{Drawing feature maps to visualize cells in 2D space with superimposed features}
\usage{
feature_score_visual(
  clusting_result,
  feature_using = NULL,
  feature_scores,
  method = "umap",
  colormap = "GnBu",
  color_direction = 1,
  nCol = NULL,
  show.axes = T,
  cell_size = 0.3,
  show_legend = T,
  show_legend_combined = F
)
}
\arguments{
\item{clusting_result}{UMAP/t-SNE results}

\item{feature_using}{Eigenvector}

\item{feature_scores}{Matrix containing feature scores}

\item{method}{Dimensionality reduction methods, e.g. tsne, umap}

\item{colormap}{RColorbrewer color palette to use}

\item{color_direction}{Sets the order of the colors in the scale. If 1 (default), the colors will be in the order of the output of RColorBrewer::brewer.pal(). If -1, the colors are in reverse order}

\item{nCol}{Drawing columns}

\item{show.axes}{Whether to display axes}

\item{cell_size}{Size of the point}

\item{show_legend}{Whether to display a single legend}

\item{show_legend_combined}{Whether to display only one legend}
}
\description{
Drawing feature maps to visualize cells in 2D space with superimposed features
}
