% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{FeatureHeatmap}
\alias{FeatureHeatmap}
\title{Generation of heat maps of differentially characterized expression of different cell populations}
\usage{
FeatureHeatmap(
  element,
  clusting_results,
  assay,
  feature_using,
  class = "celltype",
  color = NULL
)
}
\arguments{
\item{element}{Eigenexpression matrix}

\item{clusting_results}{UMAP/t-SNE results}

\item{assay}{Define the assay to be displayed, e.g., assay = “RNA”}

\item{feature_using}{Eigenvector}

\item{class}{Basis for clustering}

\item{color}{Color}
}
\description{
Generation of heat maps of differentially characterized expression of different cell populations
}
