% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{Fold_RE_TG_multiAdjust}
\alias{Fold_RE_TG_multiAdjust}
\title{Fold_RE_TG_multiAdjust}
\usage{
Fold_RE_TG_multiAdjust(X1, X2, GeneLoc, PeakLoc)
}
\arguments{
\item{X1}{scRNA-seq data matrix}

\item{X2}{scATAC-seq data matrix}

\item{GeneLoc}{Location of genes on chromosomes}

\item{PeakLoc}{Location of peaks on chromosomes}
}
\description{
Fold_RE_TG_multiAdjust
}
