% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{jaccard_matrix}
\alias{jaccard_matrix}
\title{Generate an initialized jaccard similarity matrix}
\usage{
jaccard_matrix(matrixmtx)
}
\arguments{
\item{matrixmtx}{Data matrix}
}
\description{
Generate an initialized jaccard similarity matrix
}
