% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{LRtest}
\alias{LRtest}
\title{Likelihood ratio test for bimodal distribution (fully merged version)}
\usage{
LRtest(x, y, xmin = 0)
}
\arguments{
\item{x}{Numeric vector (group 1)}

\item{y}{Numeric vector (group 2)}

\item{xmin}{Threshold separating low/high modes}
}
\description{
Likelihood ratio test for bimodal distribution (fully merged version)
}
