% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{metanet}
\alias{metanet}
\title{Construct a new network matrix by transforming the input matrix using sparse matrices}
\usage{
metanet(A, B)
}
\arguments{
\item{A}{Matrix}

\item{B}{Index vector}
}
\description{
Construct a new network matrix by transforming the input matrix using sparse matrices
}
