% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCNMF.R
\name{scClustering}
\alias{scClustering}
\title{Single-cell clustering based integration results}
\usage{
scClustering(H, k = 20)
}
\arguments{
\item{H}{Matrix for integrating results}

\item{k}{Number of nearest neighbors for KNN graph}
}
\description{
Single-cell clustering based integration results
}
